+++ Source/Engine/Material/UniformData.h
@@ -9,8 +9,8 @@ struct GlobalUBO // Slot 0, Binding 0
 {
     alignas(4) float time;
     alignas(4) float dt;
-    alignas(8) glm::ivec2 resolution;
-    alignas(8) glm::ivec2 mouse;
+    alignas(8) glm::vec2 resolution;
+    alignas(8) glm::vec2 mouse;
 
     alignas(16) glm::mat4 view;
     alignas(16) glm::mat4 projection;
